backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic[email]
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
httpx==0.26.0
feedparser==6.0.10
python-multipart==0.0.6
